Split IRC-specific code from classes.Protocol into a new IRCCommonProtocol (#454)
